Refactor to break objects out of builder and assembler.
- Moves objects into object specific files.
- Keeps inline with upstream Nengo changes.
- Remove unnecessary code, neaten the remove pass node functionality
  - Requires ctn-waterloo/nengo-temp#446
