+++ nengo_spinnaker/utils/__init__.py
@@ -8,15 +8,3 @@ from . import keyspaces
 from . import nodes
 from . import probes
 from . import vertices
-
-
-def get_connection_width(connection):
-    """Return the width of a Connection."""
-    if isinstance(connection, int):
-        return connection
-    elif isinstance(connection.post, nengo.Ensemble):
-        return connection.post.dimensions
-    elif isinstance(connection.post, nengo.Node):
-        return connection.post.size_in
-    else:
-        raise NotImplementedError
